+++ server/api.js
@@ -7,7 +7,7 @@ import config from "./utils/config";
 import { roleBasedAuth } from "./utils/middleware";
 import fetchActivity from "./controller/fetchActivity";
 import fetchReadme from "./controller/fetchReadme";
-
+import infoRouter from "./routes/getInfoRouter";
 const router = Router();
 
 router.get("/", (_, res) => {
@@ -199,5 +199,5 @@ router.put(
 		}
 	}
 );
-
+router.use("/info", infoRouter);
 export default router;
